fix(omo): avoid duplicate Orca agent roles
